labelling/requirements.txt
pandas
google-generativeai
openai
dotenv
sklearn
